requirements.txt
absl-py==0.10.0
imgaug==0.4
tensorflow-gpu==2.1.0
opencv-python==4.1.1.26
lxml==4.6.1
tqdm==4.51.0

-e .
